requirements.txt
flask==3.0.0
anthropic==0.39.0
requests==2.31.0
python-dotenv==1.0.0
markupsafe==2.1.3
